Leaflet triggers the popup even if the mouse is over the popup. Place actions in the correct place.
