+++ web/client/actions/__tests__/mousePosition-test.js
@@ -11,13 +11,9 @@ var {
     CHANGE_MOUSE_POSITION,
     CHANGE_MOUSE_POSITION_CRS,
     CHANGE_MOUSE_POSITION_STATE,
-    MOUSE_MOVE_MAP_EVENT,
-    MOUSE_OUT,
     changeMousePosition,
     changeMousePositionCrs,
-    changeMousePositionState,
-    mouseMoveMapEvent,
-    mouseOut
+    changeMousePositionState
 } = require('../mousePosition');
 
 describe('Test correctness of mausePosition actions', () => {
@@ -54,18 +50,6 @@ describe('Test correctness of mausePosition actions', () => {
         expect(retval.type).toBe(CHANGE_MOUSE_POSITION_STATE);
         expect(retval.enabled).toExist();
         expect(retval.enabled).toBe(true);
-    });
-    it('mouseMoveMapEvent', () => {
-        const position = {lat: 100, lng: 200};
-        const retval = mouseMoveMapEvent(position);
-        expect(retval).toExist();
-        expect(retval.type).toEqual(MOUSE_MOVE_MAP_EVENT);
-        expect(retval.position).toEqual(position);
-    });
-    it('mouseOut', () => {
-        const retval = mouseOut();
-        expect(retval).toExist();
-        expect(retval.type).toEqual(MOUSE_OUT);
     });
 
 });
